test: custom schema matcher
