+++ test/user.test.ts
@@ -1,18 +1,9 @@
 import { describe, expect } from '@jest/globals'
-import type { Config } from 'ts-json-schema-generator'
-import { createGenerator } from 'ts-json-schema-generator'
 import { Skolengo } from '../src/index'
-import Ajv from 'ajv'
+import './common'
 
 const SKOLENGO_TOKENSET = process.env.SKOLENGO_TOKENSET
 const describeIfLoggedIn = SKOLENGO_TOKENSET !== undefined ? describe : describe.skip
-
-const ajv = new Ajv()
-
-const ajvConfig: Config = {
-  path: 'src/**/*.ts',
-  tsconfig: 'tsconfig.json'
-}
 
 /**
  * Tests unitaires des endpoints qui nécessitent une authentification
@@ -31,36 +22,24 @@ describeIfLoggedIn('Test Skolengo API types - User logged in', () => {
   it('should getUserInfo return User type', async () => {
     const type = 'User'
     const response = await user.getUserInfo()
-    const schema = createGenerator({ ...ajvConfig, type }).createSchema(type)
 
-    const result = ajv.validate(schema, response)
-    if (!result) console.error(ajv.errors)
-
-    expect(result).toBe(true)
+    expect(response).toMatchSchema(type)
   })
 
   it('should getEvaluationSettings return EvaluationSettings type', async () => {
     if (!userPermissions.includes('READ_EVALUATIONS')) return
     const type = 'EvaluationSettings'
     const response = await user.getEvaluationSettings()
-    const schema = createGenerator({ ...ajvConfig, type }).createSchema(type)
 
     for (const evaluationSettings of response) {
-      const result = ajv.validate(schema, evaluationSettings)
-      if (!result) console.error(ajv.errors)
-
-      expect(result).toBe(true)
+      expect(evaluationSettings).toMatchSchema(type)
     }
   })
 
   it('should getUsersMailSettings return UsersMailSettings type', async () => {
     const type = 'UsersMailSettings'
     const response = await user.getUsersMailSettings()
-    const schema = createGenerator({ ...ajvConfig, type }).createSchema(type)
 
-    const result = ajv.validate(schema, response)
-    if (!result) console.error(ajv.errors)
-
-    expect(result).toBe(true)
+    expect(response).toMatchSchema(type)
   })
 })
